refactor: separate database itens into knew files
